requirements.txt
algorithms==0.1.4
beautifulsoup4==4.12.2
colorama==0.4.4
dominate==2.4.0
eventlet==0.33.3
Flask==2.1.0
Flask_Assets==2.0
Flask_Cors==3.0.10
Flask_SocketIO==5.3.3
folium==0.12.1.post1
geojson==2.5.0
geopandas==0.6.1
imutils==0.5.4
matplotlib==3.3.1
networkx==2.5
numpy==1.19.1
opencv_python==4.4.0.44
overpy==0.4
Pillow==10.1.0
PyYAML==6.0.1
rdp==0.8
requests==2.28.2
scikit_image==0.17.2
scipy==1.7.3
setuptools==50.3.0.post20201006
Shapely==1.6.4.post2
skimage==0.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
visdom==0.2.4
wandb==0.16.1
webcolors==1.13
